requirements.txt
streamlit
tensorflow
Pillow
rembg
numpy